calculator-service/requirements.txt
fastapi
uvicorn
prometheus-client
aioredis
python-dotenv